+++ terraform/aws/analytical-platform-data-production/kops/terraform.tf
@@ -10,7 +10,7 @@ terraform {
   required_providers {
     aws = {
       source  = "hashicorp/aws"
-      version = "5.15.0"
+      version = "5.17.0"
     }
   }
   required_version = "~> 1.5"