+++ terraform/aws/analytical-platform-development/cluster/terraform.tf
@@ -10,11 +10,11 @@ terraform {
   required_providers {
     auth0 = {
       source  = "auth0/auth0"
-      version = "0.50.1"
+      version = "0.50.2"
     }
     aws = {
       source  = "hashicorp/aws"
-      version = "5.15.0"
+      version = "5.17.0"
     }
     kubernetes = {
       source  = "hashicorp/kubernetes"