+++ terraform/aws/analytical-platform-development/open-metadata/vpc.tf
@@ -8,7 +8,7 @@ module "vpc" {
   #checkov:skip=CKV2_AWS_19:This is not production infrastructure
 
   source  = "terraform-aws-modules/vpc/aws"
-  version = "5.1.1"
+  version = "5.1.2"
 
   name = "open-metadata"