+++ containers/daap-get-glue-metadata/CHANGELOG.md
@@ -8,6 +8,20 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.1.2]
+
+### Changed
+
+- Bump base image to pick up logging changes
+
+## [1.1.1]
+
+### Changed
+
+- Return a 400 error if required parameters are missing
+- Return a 404 error if the table does not exist in the Glue catalog
+- Ensure Content-Type header is set to 'application/json'.
+
 ## [1.1.0]
 
 ### Changed